requirements.txt
fastapi
uvicorn
pydantic
PyJWT
requests
werkzeug
SQLAlchemy